requirements.txt
aiohttp<3.7.0
discord.py
mcstatus
pyyaml
jishaku
requests
arrow
datetime
idna<3
certifi>=2017.4.17
classy-json
discord-pretty-help
feedparser
pytz
beautifulsoup4
aiodns
discord-py-slash-command